+++ packages/moonstone/ExpandableItem/ExpandableItem.js
@@ -387,6 +387,7 @@ const ExpandableItemBase = kind({
 				aria-disabled={disabled}
 				disabled={disabled}
 				ref={setContainerNode}
+				spotlightDisabled={spotlightDisabled || disabled}
 			>
 				<LabeledItem
 					{...ariaProps}