+++ packages/moonstone/ExpandableList/tests/ExpandableList-specs.js
@@ -4,7 +4,7 @@ import {ExpandableListBase} from '../ExpandableList';
 
 describe('ExpandableList', () => {
 	describe('#aria-multiselectable', () => {
-		it('should be true when select is multiple', function () {
+		test('should be true when select is multiple', () => {
 			const expandableList = mount(
 				<ExpandableListBase title="Item" select="multiple">
 					{['option1', 'option2', 'option3']}
@@ -13,11 +13,11 @@ describe('ExpandableList', () => {
 
 			const expected = true;
 			const actual = expandableList.find('ExpandableItem').prop('aria-multiselectable');
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 	});
 
-	it('should update when children are added', function () {
+	test('should update when children are added', () => {
 		const children = ['option1', 'option2', 'option3'];
 
 		const expandableList = mount(
@@ -32,6 +32,36 @@ describe('ExpandableList', () => {
 		const expected = 5;
 		const actual = expandableList.find('GroupItem').length;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
+	});
+
+	test('should set "data-webos-voice-disabled" to LabeledItem when voice control is disabled', () => {
+		const children = ['option1', 'option2', 'option3'];
+
+		const expandableList = mount(
+			<ExpandableListBase data-webos-voice-disabled title="Item" open>
+				{children}
+			</ExpandableListBase>
+		);
+
+		const expected = true;
+		const actual = expandableList.find('LabeledItem').prop('data-webos-voice-disabled');
+
+		expect(actual).toBe(expected);
+	});
+
+	test('should set "data-webos-voice-disabled" to child item when voice control is disabled', () => {
+		const children = ['option1', 'option2', 'option3'];
+
+		const expandableList = mount(
+			<ExpandableListBase data-webos-voice-disabled title="Item" open>
+				{children}
+			</ExpandableListBase>
+		);
+
+		const expected = true;
+		const actual = expandableList.find('GroupItem').first().prop('data-webos-voice-disabled');
+
+		expect(actual).toBe(expected);
 	});
 });