+++ packages/moonstone/RadioItem/RadioItem.js
@@ -1,5 +1,5 @@
 /**
- * Provides a Moonstone-themed Item component and interactive togglable radio icon.
+ * Provides a Moonstone-themed Item component and interactive toggleable radio icon.
  *
  * @example
  * <RadioItem>Item</RadioItem>
@@ -24,6 +24,7 @@ import componentCss from './RadioItem.less';
  * @class RadioItem
  * @memberof moonstone/RadioItem
  * @extends moonstone/ToggleItem.ToggleItem
+ * @omit iconComponent
  * @ui
  * @public
  */