+++ packages/sampler/stories/qa/Marquee.js
@@ -1,7 +1,12 @@
-import {Marquee, MarqueeController} from '@enact/moonstone/Marquee';
+import kind from '@enact/core/kind';
+import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
+import Button from '@enact/moonstone/Button';
+import Divider from '@enact/moonstone/Divider';
+import Icon from '@enact/moonstone/Icon';
 import Item from '@enact/moonstone/Item';
-import ri from '@enact/ui/resolution';
+import {Marquee, MarqueeController} from '@enact/moonstone/Marquee';
 import Spottable from '@enact/spotlight/Spottable';
+import ri from '@enact/ui/resolution';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 
@@ -28,6 +33,81 @@ const RTL = [
 ];
 
 const disabledDisclaimer = (disabled) => (disabled ? <p style={{fontSize: '70%', fontStyle: 'italic'}}><sup>*</sup>Marquee does not visually respond to <code>disabled</code> state.</p> : <p />);
+
+const MarqueI18nSamples = I18nContextDecorator({updateLocaleProp: 'updateLocale'}, kind({
+	name: 'I18nPanel',
+
+	handlers: {
+		updateLocale: (ev, {updateLocale}) => updateLocale('ar-SA')
+	},
+
+	render: ({updateLocale}) => (
+		<div>
+			<Divider>Remeasure marquee when locale change causes a font change with different metrics</Divider>
+			<Button onClick={updateLocale}>change locale</Button>
+		</div>
+	)
+}));
+
+// eslint-disable-next-line enact/prop-types
+const CustomItemBase = ({children, marqueeRef, ...rest}) => (
+	<div {...rest} style={{display: 'flex', width: 300, alignItems: 'center'}}>
+		<Icon>flag</Icon>
+		<Marquee ref={marqueeRef} style={{flex: 1, overflow: 'hidden'}}>{children}</Marquee>
+		<Icon>trash</Icon>
+	</div>
+);
+
+const CustomItem = Spottable(MarqueeController(
+	{marqueeOnFocus: true},
+	CustomItemBase
+));
+
+class MarqueeWithShortContent extends React.Component {
+	constructor (props) {
+		super(props);
+
+		this.state = {
+			long: false,
+			scrollWidth: null,
+			width: null
+		};
+	}
+
+	componentDidMount () {
+		this.updateSizeInfo();
+	}
+
+	componentDidUpdate () {
+		this.updateSizeInfo();
+	}
+
+	updateSizeInfo = () => {
+		if (this.node.scrollWidth !== this.state.scrollWidth) {
+			this.setState({
+				scrollWidth: this.node.scrollWidth,
+				width: this.node.getBoundingClientRect().width
+			});
+		}
+	}
+
+	handleClick = () => {
+		this.setState(prevState => ({long: !prevState.long}));
+	}
+
+	getRef = ref => {
+		this.node = ref && ref.node;
+	}
+
+	render () {
+		return (
+			<div>
+				scrollWidth: {this.state.scrollWidth} width: {this.state.width}
+				<CustomItem marqueeRef={this.getRef} onClick={this.handleClick}>{this.state.long ? 'Very very very very very very very very very long text' : 'text'}</CustomItem>
+			</div>
+		);
+	}
+}
 
 storiesOf('Marquee', module)
 	.add(
@@ -156,5 +236,21 @@ storiesOf('Marquee', module)
 					{LTR[0]}
 				</Marquee>
 			</SpottableDiv>
+		)
+	)
+
+	.add(
+		'I18n',
+		() => (
+			<MarqueI18nSamples />
+		)
+	)
+
+	.add(
+		'with Short Content',
+		() => (
+			<div>
+				<MarqueeWithShortContent />
+			</div>
 		)
 	);