+++ packages/ui/Marquee/MarqueeDecorator.js
@@ -1,9 +1,9 @@
 import direction from 'direction';
+import {on, off} from '@enact/core/dispatcher';
 import {forward} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
-import {contextTypes as stateContextTypes} from '@enact/core/internal/PubSub';
 import {is} from '@enact/core/keymap';
-import {on, off} from '@enact/core/dispatcher';
+import {contextTypes as stateContextTypes} from '@enact/core/internal/PubSub';
 import {Job} from '@enact/core/util';
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -69,7 +69,7 @@ const defaultConfig = {
 	* Invalidate the distance if any property (like 'inline') changes.
 	* Expects an array of props which on change trigger invalidateMetrics.
 	*
-	* @type {Array}
+	* @type {String[]}
 	* @default ['remeasure']
 	* @memberof ui/Marquee.MarqueeDecorator.defaultConfig
 	*/
@@ -98,10 +98,10 @@ const defaultConfig = {
 	marqueeDirection: (str) => direction(str) === 'rtl' ? 'rtl' : 'ltr'
 };
 
-/**
+/*
  * Checks whether any of the invalidateProps has changed or not
  *
- * @param {Array} propList An array of invalidateProps
+ * @param {String[]} propList An array of invalidateProps
  * @param {Object} prev Previous props
  * @param {Object} next Next props
  * @returns {Boolean} `true` if any of the props changed
@@ -139,6 +139,18 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const forwardFocus = forward(focus);
 	const forwardEnter = forward(enter);
 	const forwardLeave = forward(leave);
+
+	const determineTextDirection = (node, rtl, forceDirection) => {
+		// Text directionality is a function of locale direction (rtl), content (node.textContent),
+		// and props (forceDirection) in increasing order of significance.
+		if (forceDirection) {
+			rtl = forceDirection === 'rtl';
+		} else if (node) {
+			rtl = marqueeDirection(node.textContent) === 'rtl';
+		}
+
+		return rtl;
+	};
 
 	return class extends React.Component {
 		static displayName = 'ui:MarqueeDecorator'
@@ -183,6 +195,15 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 */
 			forceDirection: PropTypes.oneOf(['rtl', 'ltr']),
+
+			/**
+			 * The current locale as a
+			 * {@link https://tools.ietf.org/html/rfc5646|BCP 47 language tag}.
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			locale: PropTypes.string,
 
 			/**
 			 * Number of milliseconds to wait before starting marquee when `marqueeOn` is `'hover'` or
@@ -252,7 +273,15 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 *
 			 * @private
 			 */
-			remeasure: PropTypes.any
+			remeasure: PropTypes.any,
+
+			/**
+			 * Indicates the text directionality of the current locale is right-to-left
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			rtl: PropTypes.bool
 		}
 
 		static contextTypes = {
@@ -274,20 +303,18 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				animating: false,
 				overflow: 'ellipsis',
 				promoted: false,
-				rtl: false
+				rtl: determineTextDirection(null, props.rtl, props.forceDirection)
 			};
-			this.textDirectionValidated = false;
 			this.sync = false;
 			this.forceRestartMarquee = false;
 			this.timerState = TimerState.CLEAR;
-
-			this.invalidateMetrics();
+			this.distance = null;
+			this.contentFits = false;
 		}
 
 		componentWillMount () {
 			if (this.context.Subscriber) {
 				this.context.Subscriber.subscribe('resize', this.handleResize);
-				this.context.Subscriber.subscribe('i18n', this.handleLocaleChange);
 			}
 		}
 
@@ -300,7 +327,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				});
 			}
 
-			this.validateTextDirection(this.props);
+			this.validateTextDirection();
 			if (this.props.marqueeOn === 'render') {
 				this.startAnimation(this.props.marqueeOnRenderDelay);
 			}
@@ -308,9 +335,13 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		componentWillReceiveProps (next) {
-			const {forceDirection, marqueeOn, marqueeDisabled, marqueeSpeed} = this.props;
-			this.validateTextDirection(next);
-			if (!shallowEqual(this.props.children, next.children) || (invalidateProps && didPropChange(invalidateProps, this.props, next))) {
+			const {forceDirection, locale, marqueeOn, marqueeDisabled, marqueeSpeed, rtl} = this.props;
+			if (
+				locale !== next.locale ||
+				rtl !== next.rtl ||
+				!shallowEqual(this.props.children, next.children) ||
+				(invalidateProps && didPropChange(invalidateProps, this.props, next))
+			) {
 				// restart marqueeOn="render" marquees or synced marquees that were animating
 				this.forceRestartMarquee = next.marqueeOn === 'render' || (
 					this.sync && (this.state.animating || this.timerState > TimerState.CLEAR)
@@ -318,7 +349,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 				this.invalidateMetrics();
 				this.cancelAnimation();
-				this.textDirectionValidated = false;
 			} else if (
 				next.marqueeOn !== marqueeOn ||
 				next.marqueeDisabled !== marqueeDisabled ||
@@ -332,22 +362,14 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		shouldComponentUpdate (nextProps, nextState) {
-			const {overflow, ...rest} = this.state;
-			const {overflow: nextOverflow, ...nextRest} = nextState;
-
 			return (
-				!shallowEqual(rest, nextRest) ||
-				!shallowEqual(this.props, nextProps) ||
-				(overflow !== 'ellipsis' && nextOverflow !== 'clip')
+				!shallowEqual(this.state, nextState) ||
+				!shallowEqual(this.props, nextProps)
 			);
 		}
 
 		componentDidUpdate () {
-			// if text directionality was invalidated by a prop change, we need to revalidate now
-			// potentially causing a re-render if rtl changes
-			if (this.textDirectionValidated === false) {
-				this.validateTextDirection(this.props);
-			}
+			this.validateTextDirection();
 			if (this.shouldStartMarquee()) {
 				this.tryStartingAnimation(this.props.marqueeOn === 'render' ? this.props.marqueeOnRenderDelay : this.props.marqueeDelay);
 			}
@@ -365,7 +387,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			if (this.context.Subscriber) {
 				this.context.Subscriber.unsubscribe('resize', this.handleResize);
-				this.context.Subscriber.unsubscribe('i18n', this.handleLocaleChange);
 			}
 			off('keydown', this.handlePointerHide);
 		}
@@ -447,6 +468,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.distance = null;
 			// Assume the marquee does not fit until calculations show otherwise
 			this.contentFits = false;
+
+			this.setState(state => {
+				return state.overflow === 'ellipsis' ? null : {overflow: 'ellipsis'};
+			});
 		}
 
 		/*
@@ -462,11 +487,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				this.distance = this.calculateDistance(node);
 				this.contentFits = !this.shouldAnimate(this.distance);
 
-				// TODO: Replace with functional setState with React 16
 				const overflow = this.calculateTextOverflow(this.distance);
-				if (!(this.contentFits && overflow === 'clip' && this.state.overflow === 'ellipsis')) {
-					this.setState({overflow});
-				}
+				this.setState(state => {
+					return state.overflow === overflow ? null : {overflow};
+				});
 			}
 		}
 
@@ -477,7 +501,8 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 * @returns	{Number}			Distance to travel in pixels
 		 */
 		calculateDistance (node) {
-			const distance = Math.ceil(node.scrollWidth - node.clientWidth);
+			const rect = node.getBoundingClientRect();
+			const distance = Math.abs(node.scrollWidth - rect.width);
 			return distance;
 		}
 
@@ -494,7 +519,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 * @returns	{String}				text-overflow value
 		 */
 		calculateTextOverflow (distance) {
-			return distance === 0 ? 'clip' : 'ellipsis';
+			return distance < 1 ? 'clip' : 'ellipsis';
 		}
 
 		/*
@@ -504,7 +529,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 * @returns	{Boolean}				`true` if it should animated
 		 */
 		shouldAnimate (distance) {
-			return distance > 0;
+			return distance >= 1;
 		}
 
 		/*
@@ -648,11 +673,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		handleLocaleChange = ({message: {rtl}}) => {
-			this.rtlLocale = rtl;
-			this.validateTextDirection(this.props);
-		}
-
 		handleMarqueeComplete = () => {
 			this.resetAnimation();
 		}
@@ -718,24 +738,11 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.node = node;
 		}
 
-		validateTextDirection = ({forceDirection}) => {
-			// @TODO: replace with functional setState with React 16
-			// Text directionality is a function of locale (this.rtlLocale), content
-			// (this.node.textContent), and props (this.props.forceDirection) in increasing order of
-			// significance.
-			let rtl = this.rtlLocale;
-			if (forceDirection) {
-				rtl = forceDirection === 'rtl';
-			} else if (this.node) {
-				rtl = marqueeDirection(this.node.textContent) === 'rtl';
-				this.textDirectionValidated = true;
-			}
-
-			// prevent re-render when text direction matches locale direction
-			if (rtl !== this.state.rtl) {
-				// eslint-disable-next-line react/no-did-update-set-state
-				this.setState({rtl});
-			}
+		validateTextDirection () {
+			this.setState((state, props) => {
+				const rtl = determineTextDirection(this.node, props.rtl, props.forceDirection);
+				return state.rtl === rtl ? null : {rtl};
+			});
 		}
 
 		renderMarquee () {
@@ -743,7 +750,6 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				alignment,
 				children,
 				disabled,
-				forceDirection,
 				marqueeOn,
 				marqueeSpeed,
 				...rest
@@ -767,6 +773,8 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				rest[enter] = this.handleEnter;
 			}
 
+			delete rest.forceDirection;
+			delete rest.locale;
 			delete rest.marqueeCentered;
 			delete rest.marqueeDelay;
 			delete rest.marqueeDisabled;
@@ -774,11 +782,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			delete rest.marqueeResetDelay;
 			delete rest.marqueeSpeed;
 			delete rest.remeasure;
-
-			let {rtl} = this.state;
-			if (forceDirection) {
-				rtl = forceDirection === 'rtl';
-			}
+			delete rest.rtl;
 
 			return (
 				<Wrapped {...rest} onBlur={this.handleBlur} disabled={disabled}>
@@ -790,7 +794,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 						distance={this.distance}
 						onMarqueeComplete={this.handleMarqueeComplete}
 						overflow={this.state.overflow}
-						rtl={rtl}
+						rtl={this.state.rtl}
 						speed={marqueeSpeed}
 						willAnimate={this.state.promoted}
 					>
@@ -805,6 +809,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			delete props.alignment;
 			delete props.forceDirection;
+			delete props.locale;
 			delete props.marqueeCentered;
 			delete props.marqueeDelay;
 			delete props.marqueeDisabled;
@@ -813,6 +818,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			delete props.marqueeResetDelay;
 			delete props.marqueeSpeed;
 			delete props.remeasure;
+			delete props.rtl;
 
 			return <Wrapped {...props} />;
 		}