+++ packages/ui/Transition/Transition.js
@@ -36,6 +36,7 @@ const forwardOnHide = forward('onHide');
  * In general, you'll probably want to use the `Transition` instead of `TransitionBase`.
  *
  * @class TransitionBase
+ * @ui
  * @memberof ui/Transition
  * @public
  */
@@ -292,6 +293,7 @@ const TRANSITION_STATE = {
  * properties and events.
  *
  * @class Transition
+ * @ui
  * @memberof ui/Transition
  * @public
  */
@@ -436,8 +438,10 @@ class Transition extends React.Component {
 	}
 
 	componentWillReceiveProps (nextProps) {
-		if (nextProps.visible && this.state.renderState === TRANSITION_STATE.INIT) {
+		if (!this.props.visible && nextProps.visible) {
 			this.setState({
+				initialHeight: null,
+				initialWidth: null,
 				renderState: TRANSITION_STATE.MEASURE
 			});
 		}